Add test coverage for JWT issuance nad validation
